iadvogado/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
pillow
pytesseract
# openai  # Comentado - usando Llama local
# gTTS  # Comentado - usando Edge TTS
httpx
python-dotenv
supabase
sqlalchemy
# Dependências para Llama 3.1
torch
transformers
accelerate
bitsandbytes
sentencepiece
protobuf
# Edge TTS
edge-tts